+++ packages/types/CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0] - 2024-03-28
 ### Added
-- Add IBlock interface for block content of all networks, also support dictionary to be array type (#2257) 
+- Add IBlock interface for block content of all networks, also support dictionary to be array type (#2257)
 
 ## [3.3.1] - 2024-02-07
 ### Changed
@@ -220,7 +221,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support block handler
 
-[Unreleased]: https://github.com/subquery/subql/compare/types/3.3.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/types/3.4.0...HEAD
+[3.4.0]: https://github.com/subquery/subql/compare/types/3.3.1...types/3.4.0
 [3.3.1]: https://github.com/subquery/subql/compare/types/3.3.0...types/3.3.1
 [3.3.0]: https://github.com/subquery/subql/compare/types/3.2.0...types/3.3.0
 [3.2.0]: https://github.com/subquery/subql/compare/types/3.1.2...types/3.2.0